+++ editor/plugins/collision_shape_2d_editor_plugin.cpp
@@ -35,9 +35,9 @@
 #include "scene/resources/circle_shape_2d.h"
 #include "scene/resources/concave_polygon_shape_2d.h"
 #include "scene/resources/convex_polygon_shape_2d.h"
+#include "scene/resources/line_shape_2d.h"
 #include "scene/resources/rectangle_shape_2d.h"
 #include "scene/resources/segment_shape_2d.h"
-#include "scene/resources/shape_line_2d.h"
 
 Variant CollisionShape2DEditor::get_handle_value(int idx) const {
