+++ scene/3d/SCsub
@@ -7,6 +7,6 @@ if env['disable_3d']:
     env.scene_sources.append("3d/skeleton.cpp")
     env.scene_sources.append("3d/particles.cpp")
     env.scene_sources.append("3d/visual_instance.cpp")
-    env.scene_sources.append("3d/scenario_fx.cpp")
+    env.scene_sources.append("3d/world_environment.cpp")
 else:
     env.add_source_files(env.scene_sources, "*.cpp")
