+++ scene/3d/reflection_probe.h
@@ -32,7 +32,7 @@
 #define REFLECTIONPROBE_H
 
 #include "scene/3d/visual_instance.h"
-#include "scene/resources/sky_box.h"
+#include "scene/resources/sky.h"
 #include "scene/resources/texture.h"
 #include "servers/visual_server.h"
