+++ scene/gui/link_button.h
@@ -32,7 +32,7 @@
 #define LINKBUTTON_H
 
 #include "scene/gui/base_button.h"
-#include "scene/resources/bit_mask.h"
+#include "scene/resources/bit_map.h"
 
 class LinkButton : public BaseButton {
