+++ scene/gui/texture_button.h
@@ -32,7 +32,7 @@
 #define TEXTURE_BUTTON_H
 
 #include "scene/gui/base_button.h"
-#include "scene/resources/bit_mask.h"
+#include "scene/resources/bit_map.h"
 class TextureButton : public BaseButton {
 
 	GDCLASS(TextureButton, BaseButton);
